// packages/sample-atlaskit-editor-modules/atlaskit/editor-core/plugins/floating-toolbar/types.ts
export type Icon = React.ComponentType<{ label: string }>;
export type RenderOptionsProps = RenderOptionsPropsT<Command>;

export type AlignType = 'left' | 'center' | 'right';

export interface ConfirmDialogOptions {
  title?: string; // Defaults to "Are you sure?"
  message: string;
  okButtonLabel?: string; // Defaults to "OK"
  cancelButtonLabel?: string; // Defaults to "Cancel"
}

export type FloatingToolbarButton<T> = {
  id?: string;
  type: 'button';
  title: string;
  onClick: T;
  showTitle?: boolean;
  onMouseEnter?: T;
  onMouseLeave?: T;
  icon?: Icon;
  selected?: boolean;
  disabled?: boolean;
  hidden?: boolean;
  appearance?: ButtonAppearance;
  href?: string;
  target?: string;
  className?: string;
  tooltipContent?: React.ReactNode;
  testId?: string;
  hideTooltipOnClick?: boolean;
  confirmDialog?: ConfirmDialogOptions;
  // For sending data over the mobile bridge
  metadata?: { [key: string]: string };
};
// ...
export type FloatingToolbarSelect<T, V = SelectOption> = {
  id: string;
  type: 'select';
  selectType: 'list' | 'color' | 'emoji' | 'date';
  title?: string;
  options: V[];
  hidden?: boolean;
  hideExpandIcon?: boolean;
  defaultValue?: V;
  placeholder?: string;
  onChange: (selected: V) => T;
  filterOption?: ((option: V, rawInput: string) => boolean) | null;
};

export type FloatingToolbarColorPicker<T> = FloatingToolbarSelect<
  T,
  PaletteColor
> & {
  selectType: 'color';
};

export type FloatingToolbarDatePicker<T> = FloatingToolbarSelect<T, number> & {
  selectType: 'date';
  options: never[];
};

export type FloatingToolbarSeparator = {
  type: 'separator';
  hidden?: boolean;
};
// ...
export type FloatingToolbarItem<T> =
  | FloatingToolbarButton<T>
  | FloatingToolbarDropdown<T>
  | FloatingToolbarSelect<T>
  | FloatingToolbarColorPicker<T>
  | FloatingToolbarDatePicker<T>
  | FloatingToolbarInput<T>
  | FloatingToolbarCustom<T>
  | FloatingToolbarSeparator
  | FloatingToolbarExtensionsPlaceholder;
export interface FloatingToolbarConfig {
  title: string;
  /**
   * Override the DOM reference used to apply as the target for the
   * floating toolbar, if the config matches.
   *
   * By default, it will find the DOM reference of the node from the
   * head of the current selection.
   */
  getDomRef?: (view: EditorView) => HTMLElement | undefined;

  visible?: boolean;
  nodeType: NodeType | NodeType[];
  items:
    | Array<FloatingToolbarItem<Command>>
    | ((node: Node) => Array<FloatingToolbarItem<Command>>);
  align?: AlignType;
  className?: string;
  height?: number;
  width?: number;
  offset?: [number, number];
  forcePlacement?: boolean;
  onPositionCalculated?: (
    editorView: EditorView,
    nextPos: Position,
  ) => Position;
}

export type FloatingToolbarHandler = (
  state: EditorState,
  intl: InjectedIntl,
  providerFactory: ProviderFactory,
  cardOptions?: CardOptions,
) => FloatingToolbarConfig | undefined;